requirements.txt
fastapi
uvicorn
together
python-dotenv
requests
tweepy
